refactor(api): modulariza sistema de chaves e aplica SOLID
- Cria módulo independente `api/app/modules/chave`.
- Isola schemas, controller e service exclusivos para chaves de acesso.
- Move lógica de hierarquia e validação de escopo para `ChaveService`.
- Refatora `UsuarioService` para consumir o novo módulo de chaves.
- Limpa imports e remove redundâncias no `UsuarioController`.
- Atualiza plano de evolução com etapa de Inspeção de Chaves (Fase 2.5).
